requirements.txt
aiogram==3.14.0
python-dotenv==1.0.1
asyncpg==0.29.0
cryptography==42.0.8
aiohttp==3.9.0
PyJWT==2.8.0